feat: add and use constant data service
